phuoc khung

+++ server/src/main/java/com/dungcuthethao/RestController/NguoiDungRestController.java
@@ -85,8 +85,10 @@ public class NguoiDungRestController {
 	}
 	
 	@GetMapping("/tongso")
-	public int getTotal() {
-		return (int) nguoiDungRepository.count();
+	public Long getTotal() {
+		return  nguoiDungRepository.count();
 	}
+	
+	
 	
 }
